split out sensor unittests into own files
